#3 Refactor : 가격, 수량을 Long -> int 로 바꿈
