+++ src/main/java/com/jihyuk/mojong_v2/model/enums/ROLE.java
@@ -1,4 +1,4 @@
-package com.jihyuk.mojong_v2.model;
+package com.jihyuk.mojong_v2.model.enums;
 
 public enum ROLE {
     ROLE_GUEST, ROLE_STAFF, ROLE_ADMIN
